Version 3.0.1. Reduced code for buttons, final release.
Changes:
1. Added buttons class to create button graphics, tie actual buttons to a graphic, and check button presses.
2. Eliminated many display lists for buttons in favor of Buttons class.
3. Eliminated ButtonPressCheck functions in most Views in favor of Buttons class.

Things I couldn't quite get around to doing:
1. Multithreading
2. Automaton 'Stability reached' check

Finally reached the end of a great time working on a project. Was it worth the effort? Based on the very small amount of appreciation recieved and money spent printing around 80 pages of code..... not so much. Based on self satisfaction and support from friends and family? You bet your sweet fruit bread slices it was!

-George Varghese M, 27th May 2016
[The End]